Moved jwt specific code into hexagoal folder structure
