--- backend/auth/__init__.py
@@ -1,8 +1,0 @@
-from backend.auth.auth_bearer import oauth2_scheme_dep
-from backend.auth.authenticate_user import authenticate_user
-from backend.auth.constants import ACCESS_TOKEN_EXPIRE_MINUTES
-from backend.auth.create_access_token import create_access_token
-from backend.auth.current_user import currentUserDep, get_current_user
-from backend.auth.get_user import get_user
-from backend.auth.models import Token, TokenData
-from backend.auth.verify_password import verify_password
