--- backend/auth/authenticate_user.py
@@ -1,10 +1,0 @@
-from backend.auth import get_user, verify_password
-
-
-def authenticate_user(fake_db, username: str, password: str):
-    user = get_user(fake_db, username)
-    if not user:
-        return False
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
