--- backend/auth/constants.py
@@ -1,1 +1,0 @@
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
