--- backend/auth/get_user.py
@@ -1,7 +1,0 @@
-from backend.db import UserInDB
-
-
-def get_user(db, username: str):
-    if username in db:
-        user_dict = db[username]
-        return UserInDB(**user_dict)
