+++ backend/features/auth/domain/services.py
@@ -1,0 +1,80 @@
+from datetime import datetime, timedelta, timezone
+from typing import Annotated
+
+import jwt
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+from passlib.context import CryptContext
+
+from backend.common.config import Settings, settingsDep
+from backend.db import User
+from backend.features.auth.domain.models import TokenData
+from backend.features.user.repositories import get_user
+from backend.tests.mockedData import fake_users_db
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+oauth2_scheme_dep = Annotated[str, Depends(oauth2_scheme)]
+
+
+def authenticate_user(fake_db, username: str, password: str):
+    user = get_user(fake_db, username)
+    if not user:
+        return False
+    if not verify_password(password, user.hashed_password):
+        return False
+    return user
+
+
+def create_access_token(
+    settings: Settings, data: dict, expires_delta: timedelta | None = None
+) -> str:
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.now(timezone.utc) + expires_delta
+    else:
+        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, settings.secret_key_jwt, algorithm=settings.algorithm_jwt)
+    return encoded_jwt
+
+
+def fake_decode_token(token):
+    return User(username=token + "fakedecoded", email="john@example.com", full_name="John Doe")
+
+
+async def get_current_user(token: oauth2_scheme_dep, settings: settingsDep):
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, settings.secret_key_jwt, algorithms=[settings.algorithm_jwt])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        token_data = TokenData(username=username)
+    except jwt.exceptions.InvalidTokenError:
+        raise credentials_exception
+    user = get_user(fake_users_db, username=token_data.username)
+    if user is None:
+        raise credentials_exception
+    return user
+
+
+currentUserDep = Annotated[User, Depends(get_current_user)]
+
+
+async def get_current_active_user(
+    current_user: currentUserDep,
+):
+    if current_user.disabled:
+        raise HTTPException(status_code=400, detail="Inactive user")
+    return current_user
+
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+
+def verify_password(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
