+++ backend/features/user/repositories/__init__.py
@@ -1,0 +1,1 @@
+from backend.features.user.repositories.user_repository import get_user
