+++ backend/features/user/repositories/user_repository.py
@@ -1,0 +1,7 @@
+from backend.db import UserInDB
+
+
+def get_user(db, username: str):
+    if username in db:
+        user_dict = db[username]
+        return UserInDB(**user_dict)
