+++ backend/features/user/routes/__init__.py
@@ -1,0 +1,1 @@
+from backend.features.user.routes import router as user_router
