+++ backend/middlewares.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from backend.config import settings
+from backend.common.config import settings
 
 
 def register_middlewares(app: FastAPI):
