+++ backend/routes/__init__.py
@@ -1,3 +1,1 @@
-from backend.routes.example import router as example_router
 from backend.routes.items import router as items_router
-from backend.routes.user import router as user_router
