--- backend/routes/example.py
@@ -1,9 +1,0 @@
-from fastapi import APIRouter
-from fastapi.responses import JSONResponse
-
-router = APIRouter()
-
-
-@router.get("/")
-def read_root():
-    return JSONResponse(content={"message": "Welcome to FastAPI with Conda!"})
